text input in raw mode
creating our own input_line methods provides us with greater control
over the underlying types, this way it was much easier to properly
create the features we want to have.

this way it was very easy to disregard typed input while commands are
being processed. This was the desired behaviour because a single thread
runing async tasks is responsible for listening to many events that
are critical to the features this app provides, while also keeping app
overhead as low as possible.