Queue messages when splash screen is displayed
Now we properly handle printing messages to the user when the splash
screen is displayed. A String is used to collect messages via our
`ColoredFormatter` or manually through `print_during_splash` via
`Message`s.

Should be noted `print_during_splash` is only to be used for errors
encountered before tracing subscriber has been initalized. This ensures
that the non-async `Mutex` used for the queue has no contention and
won't cause the process to hang.