Start input library documentation
Removed the requirement to impl the `Executor` trait, and the helper
method that came along with the required bounds in commit 8165231. This
helper method was really not much less boilerplate for the library user
and just made it less clear how the main process loop was being handled
within the library.