Feat: Obtain api endpoints from gist
This makes it so users do not have to update to a new version of
MatchWire if any HMW/iw4 server location url changes.

Also added support to parse the HMW verion manifest via JSONPath so even
if the format of the HMW manifest changes users will still not have to
update to a new verson.

Hardcoded fallback urls are still provided in the program if the request
to the gist fails. This also makes the code cleaner by specifying all
the urls in one place.