Add duplicate server print line
The filter command now notifies if duplicate servers are found. This
makes it less confusing of why at times the found count did not match
the count added to favourites.json.

Also fixed a using a certian combination of filter flags would lead to
duplicates not properly being filtered out.