Feat: Display server browser
Added new '--stats' feature, this allows to visually see the data
MatchWire is processing. Displaying statistics for each
source included in the current query as well as a server browser table
of the filtered query results. Along with stats a new
command 'last' was added that re-prints the last filter that included
'--stats'.

'reconnect --history' was also updated to match the styling of the new
tables.

Removed a bunch of sync primitives and replaced with thread local
storage, since all state is all shared by the same main thread.

Lots of various cleanup and orginization of modules/functions/methods

Fixed a bug where non-responsive servers would still be updated in the
connection history when trying to direct connect
Fixed a parsing edge case that wouldn't correctly determine the type of
server connection
Fixed some missed logging messages on startup by initializing the logger
as early as possible