Filter non-running servers
It's possible for server owners to leave iw4m running after either their
server crashes or they close the game server. These entries are now
filtered out.

The tracking of a missing cache modification is now tracked properly, if
it is even worth propagating up the tracked modification could be
debated. Now that the global bool is stored in TLS the code would read
much cleaner if was just bitor'd directly at the spot of modicication.