Prefer iw4m resolved addrs
Thanks to RaidMax for exposing servers resolved IPv4 addrs to the iw4m
API - RaidMax/IW4M-Admin@3e186ca

This allows us to get a valid location for home hosted servers using the
loopback configuration. This way /getInfo is easily accessable to us
over IPv4 and the same addr can be used in the game client.

IPv6 addrs in HMW are not fully supported at the moment. Additional
support will be _likely_ re-introduced in an upcoming update to the
client.