Feat: Verify all HMW files
Added `--verify-all` to the `version` command, this will allow the user
to verify the sha256 hash of all mod files, PGP verification will likely
be added in a future commit.

To accomplish this we now are caching HMW's manifest and only updating
it if the `manifest_guid` does not match.

The display module was split into smaller files with the addition of a
new indicator type `ProgressBar`.

`Cache` access from inside of filter stats (server browser) display
implementation was abstracted as it is not ideal to access globals from
within a `Display` impl. This also makes our test case for it much
cleaner.