Batch region requests
The single ip location endpoint had a ratelimit that we would previously
hit whenever the cache was empty. To scale up the ammount of ips we can
retrieve location data for within a 60 second period we use the batching
endpoint that API provider supplies. This can be used to request the
locaton for 100 ips at a time, we now take advantage of this.

This endpoint also has a rate limit. Though unlikely that we would hit
the new limit of 1500 ip locations within 60 seconds, a rate limiter was
added. The new `RateLimiter` ensures the users ip will never get timed
out from sending too many requests.