PGP verify startup endpoints
We now verify the startup json that is fetched everytime MatchWire is
launched. This provides a few extra layers of security. The locations
for HMW's public key and manifest data need to pass the signature check.
Instead of comparing version strings of MatchWire the latest hash is
stored in the verified endpoint.

Since we now rely on Signature verified version data, we no longer have
backup endpoints in case of errors retrieving the startup data not
related to bad verification.